identified the root cause of the runtime warning: sometimes a centroid ends up with zero members after an update. we need to decide on and implement a way to handle such cases
